Refactor out ProvingThreadConfig to module
