+++ ccp/src/cu/proving_thread/async_/config.rs
@@ -14,10 +14,21 @@
  * limitations under the License.
  */
 
-mod config;
-mod errors;
-mod thread;
+use crate::cu::CUProverConfig;
 
-pub(crate) use errors::ProvingThreadAsyncError;
-pub(crate) use thread::ProvingThreadAsync;
-pub(crate) use thread::ProvingThreadConfig;
+#[derive(Debug, Clone)]
+pub struct ProvingThreadConfig {
+    pub hashes_per_round: usize,
+    pub async_to_sync_queue_size: usize,
+    pub sync_to_async_queue_size: usize,
+}
+
+impl ProvingThreadConfig {
+    pub fn from_cu_prover_config(cu_config: &CUProverConfig) -> Self {
+        Self {
+            hashes_per_round: cu_config.hashes_per_round,
+            async_to_sync_queue_size: cu_config.async_to_sync_queue_size,
+            sync_to_async_queue_size: cu_config.sync_to_async_queue_size,
+        }
+    }
+}
